Move last message display to its own component
